use content_for to insert the named yield block into the layout
