--- app/components/captcha_script_component.html.erb
@@ -1,3 +1,0 @@
-<% if recaptcha_script_src.present? %>
-  <%= content_tag(:script, '', src: recaptcha_script_src, async: true) %>
-<% end %>
